requirements.txt
blinker==1.9.0
click==8.2.1
dominate==2.9.1
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-flask==1.3.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.14.1
visitor==0.1.3
Werkzeug==3.1.3
